Extract TypeFormatter and LexicalScope
