+++ lib/rbs_infer/analyzer.rb
@@ -1266,6 +1266,8 @@ require_relative "inference/param_type_inferrer"
 require_relative "project/source_index"
 require_relative "project/mixin_index"
 require_relative "signatures/steep_bridge"
+require_relative "signatures/steep_bridge/lexical_scope"
+require_relative "signatures/steep_bridge/type_formatter"
 require_relative "signatures/steep_bridge/ivar_write_analyzer"
 require_relative "project/source_expanders"
 require_relative "project/self_type_annotators"
