+++ spec/lib/rbs_infer/signatures/steep_bridge_spec.rb
@@ -222,10 +222,10 @@ RSpec.describe RbsInfer::Signatures::SteepBridge, :dummy_app do
       # handling, `to_s` emits `<% Steep::AST::Types::Logic::Not %>`
       # which leaks into the generated RBS as a literal `Logic::Not`
       # string. Verify the helper collapses each Logic type to `bool`.
-      expect(bridge.send(:format_type, Steep::AST::Types::Logic::Not.instance)).to eq("bool")
-      expect(bridge.send(:format_type, Steep::AST::Types::Logic::ReceiverIsNil.instance)).to eq("bool")
-      expect(bridge.send(:format_type, Steep::AST::Types::Logic::ReceiverIsArg.instance)).to eq("bool")
-      expect(bridge.send(:format_type, Steep::AST::Types::Logic::ArgIsReceiver.instance)).to eq("bool")
+      expect( RbsInfer::Signatures::SteepBridge::TypeFormatter.format_type(Steep::AST::Types::Logic::Not.instance)).to eq("bool")
+      expect(RbsInfer::Signatures::SteepBridge::TypeFormatter.format_type(Steep::AST::Types::Logic::ReceiverIsNil.instance)).to eq("bool")
+      expect(RbsInfer::Signatures::SteepBridge::TypeFormatter.format_type(Steep::AST::Types::Logic::ReceiverIsArg.instance)).to eq("bool")
+      expect(RbsInfer::Signatures::SteepBridge::TypeFormatter.format_type(Steep::AST::Types::Logic::ArgIsReceiver.instance)).to eq("bool")
     end
 
     it "resolves chained method return types" do
